requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-SocketIO==5.3.6
Flask-Babel==4.0.0
Flask-Compress==1.14
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-Assets==2.1.0
Flask-HTTPAuth==4.8.0
python-dotenv==1.0.0
SQLAlchemy==2.0.27
psycopg2==2.9.9
psycopg2-binary==2.9.9
Werkzeug==3.0.1
WTForms==3.1.2
email-validator==2.1.0.post1
Pillow==10.1.0
python-socketio==5.11.1
redis==5.0.1
requests==2.31.0
bcrypt==4.1.1
PyJWT==2.8.0
gunicorn==21.2.0
eventlet==0.35.2
python-dateutil==2.8.2
Markdown==3.5.1
bleach==6.1.0
pytz==2023.3.post1
Babel==2.13.1
webassets==2.0
cssmin==0.2.0
jsmin==3.0.1
rq==1.15.1
alembic==1.13.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
python-engineio==4.9.0
six==1.16.0
typing_extensions==4.9.0
pyotp==2.8.0
qrcode==7.4.2
whitenoise==6.6.0 